API

AbstractMCMC defines an interface for sampling Markov chains.

Sampling a single chain

StatsBase.sample — Method

sample([rng, ]model, sampler, N; kwargs...)

Return N samples from the model with the Markov chain Monte Carlo sampler.

StatsBase.sample — Method

sample([rng, ]model, sampler, isdone; kwargs...)

Sample from the model with the Markov chain Monte Carlo sampler until a convergence criterion isdone returns true, and return the samples.

The function isdone has the signature

isdone(rng, model, sampler, samples, iteration; kwargs...)

and should return true when sampling should end, and false otherwise.

Iterator

AbstractMCMC.steps — Method

steps([rng, ]model, sampler; kwargs...)

Create an iterator that returns samples from the model with the Markov chain Monte Carlo sampler.

Examples

julia> struct MyModel <: AbstractMCMC.AbstractModel end

julia> struct MySampler <: AbstractMCMC.AbstractSampler end

julia> function AbstractMCMC.step(rng, ::MyModel, ::MySampler, state=nothing; kwargs...)
           # all samples are zero
           return 0.0, state
       end

julia> iterator = steps(MyModel(), MySampler());

julia> collect(Iterators.take(iterator, 10)) == zeros(10)
true

Transducer

AbstractMCMC.Sample — Method

Sample([rng, ]model, sampler; kwargs...)

Create a transducer that returns samples from the model with the Markov chain Monte Carlo sampler.

Examples

julia> struct MyModel <: AbstractMCMC.AbstractModel end

julia> struct MySampler <: AbstractMCMC.AbstractSampler end

julia> function AbstractMCMC.step(rng, ::MyModel, ::MySampler, state=nothing; kwargs...)
           # all samples are zero
           return 0.0, state
       end

julia> transducer = Sample(MyModel(), MySampler());

julia> collect(transducer(1:10)) == zeros(10)
true

Sampling multiple chains in parallel

StatsBase.sample — Method

sample([rng, ]model, sampler, parallel, N, nchains; kwargs...)

Sample nchains Monte Carlo Markov chains from the model with the sampler in parallel using the parallel algorithm, and combine them into a single chain.

Two algorithms are provided for parallel sampling with multiple threads and multiple processes, and one allows for the user to sample multiple chains in serial (no parallelization):

AbstractMCMC.MCMCThreads — Type

MCMCThreads

The MCMCThreads algorithm allows users to sample MCMC chains in parallel using multiple threads.

AbstractMCMC.MCMCDistributed — Type

MCMCDistributed

The MCMCDistributed algorithm allows users to sample MCMC chains in parallel using multiple processes.

AbstractMCMC.MCMCSerial — Type

MCMCSerial

The MCMCSerial algorithm allows users to sample serially, with no thread or process parallelism.

Common keyword arguments

Common keyword arguments for regular and parallel sampling (not supported by the iterator and transducer) are:

progress (default: AbstractMCMC.PROGRESS[] which is true initially): toggles progress logging

chain_type (default: Any): determines the type of the returned chain

callback (default: nothing): if callback !== nothing, then callback(rng, model, sampler, sample, iteration) is called after every sampling step, where sample is the most recent sample of the Markov chain and iteration is the current iteration

discard_initial (default: 0): number of initial samples that are discarded

thinning (default: 1): factor by which to thin samples.

Progress logging can be enabled and disabled globally with AbstractMCMC.setprogress!(progress).

AbstractMCMC.setprogress! — Function

setprogress!(progress::Bool)

Enable progress logging globally if progress is true, and disable it otherwise.

Chains

The chain_type keyword argument allows to set the type of the returned chain. A common choice is to return chains of type Chains from MCMCChains.jl.

AbstractMCMC defines the abstract type AbstractChains for Markov chains.

AbstractMCMC.AbstractChains — Type

AbstractChains

AbstractChains is an abstract type for an object that stores parameter samples generated through a MCMC process.

For chains of this type, AbstractMCMC defines the following two methods.

AbstractMCMC.chainscat — Function

chainscat(c::AbstractChains...)

Concatenate multiple chains.

By default, the chains are concatenated along the third dimension by calling cat(c...; dims=3).

AbstractMCMC.chainsstack — Function

chainsstack(c::AbstractVector)

Stack chains in c.

By default, the vector of chains is returned unmodified. If eltype(c) <: AbstractChains, then reduce(chainscat, c) is called.

Interacting with states of samplers

To make it a bit easier to interact with some arbitrary sampler state, we encourage implementations of AbstractSampler to implement the following methods:

AbstractMCMC.realize — Function

realize(transition)

Return the realization of the random variables present in transition.

AbstractMCMC.realize!! — Function

realize!!(state, realization)

Update the realization of the state with realization and return it.

If state can be updated in-place, it is expected that this function returns state with updated realize. Otherwise a new state object with the new realization is returned.

and optionally

AbstractMCMC.updatestate!! — Method

updatestate!!(state, transition_prev[, state_prev])

Return new instance of state using information from transition_prev and, optionally, state_prev.

Defaults to realize!!(state, realize(transition_prev)).

These methods can also be useful for implementing samplers which wraps some inner samplers, e.g. a mixture of samplers.

Example: MixtureSampler

In a MixtureSampler we need two things:

components: collection of samplers.

weights: collection of weights representing the probability of chosing the corresponding sampler.

struct MixtureSampler{W,C} <: AbstractMCMC.AbstractSampler
    components::C
    weights::W
end

To implement the state, we need to keep track of a couple of things:

index: the index of the sampler used in this step.

transition: the transition resulting from this step.

states: the current states of all the components.

Two aspects of this might seem a bit strange:

We need to keep track of the states of all components rather than just the state for the sampler we used previously.

We need to put the transition from the step into the state.

The reason for (1) is that lots of samplers keep track of more than just the previous realizations of the variables, e.g. in AdvancedHMC.jl we keep track of the momentum used, the metric used, etc.

For (2) the reason is similar: some samplers might keep track of the variables in the state differently, e.g. you might have a sampler which is independent of the current realizations and the state is simply nothing.

Hence, we need the transition, which should always contain the realizations, to make sure we can resume from the same point in the space in the next step.

struct MixtureState{T,S}
    index::Int
    transition::T
    states::S
end

The step for a MixtureSampler is defined by the following generative process

\[\begin{aligned} i &\sim \mathrm{Categorical}(w_1, \dots, w_k) \\ X_t &\sim \mathcal{K}_i(\cdot \mid X_{t - 1}) \end{aligned}\]

where $\mathcal{K}_i$ denotes the i-th kernel/sampler, and $w_i$ denotes the weight/probability of choosing the i-th sampler. AbstractMCMC.updatestate!! comes into play in defining/computing $\mathcal{K}_i(\cdot \mid X_{t - 1})$ since $X_{t - 1}$ could be coming from a different sampler.

If we let state be the current MixtureState, i the current component, and i_prev is the previous component we sampled from, then this translates into the following piece of code:

# Update the corresponding state, i.e. `state.states[i]`, using
# the state and transition from the previous iteration.
state_current = AbstractMCMC.updatestate!!(
    state.states[i], state.states[i_prev], state.transition
)

# Take a `step` for this sampler using the updated state.
transition, state_current = AbstractMCMC.step(
    rng, model, sampler_current, sampler_state;
    kwargs...
)

The full AbstractMCMC.step implementation would then be something like:

function AbstractMCMC.step(rng, model::AbstractMCMC.AbstractModel, sampler::MixtureSampler, state; kwargs...)
    # Sample the component to use in this `step`.
    i = rand(Categorical(sampler.weights))
    sampler_current = sampler.components[i]

    # Update the corresponding state, i.e. `state.states[i]`, using
    # the state and transition from the previous iteration.
    i_prev = state.index
    state_current = AbstractMCMC.updatestate!!(
        state.states[i], state.states[i_prev], state.transition
    )

    # Take a `step` for this sampler using the updated state.
    transition, state_current = AbstractMCMC.step(
        rng, model, sampler_current, state_current;
        kwargs...
    )

    # Create the new states.
    # NOTE: Code below will result in `states_new` being a `Vector`.
    # If we wanted to allow usage of alternative containers, e.g. `Tuple`,
    # it would be better to use something like `@set states[i] = state_current`
    # where `@set` is from Setfield.jl.
    states_new = map(1:length(state.states)) do j
        if j == i
            # Replace the i-th state with the new one.
            state_current
        else
            # Otherwise we just carry over the previous ones.
            state.states[j]
        end
    end

    # Create the new `MixtureState`.
    state_new = MixtureState(i, transition, states_new)

    return transition, state_new
end

And for the initial AbstractMCMC.step we have:

function AbstractMCMC.step(rng, model::AbstractMCMC.AbstractModel, sampler::MixtureSampler; kwargs...)
    # Initialize every state.
    transitions_and_states = map(sampler.components) do spl
        AbstractMCMC.step(rng, model, spl; kwargs...)
    end

    # Sample the component to use this `step`.
    i = rand(Categorical(sampler.weights))
    # Extract the corresponding transition.
    transition = first(transitions_and_states[i])
    # Extract states.
    states = map(last, transitions_and_states)
    # Create new `MixtureState`.
    state = MixtureState(i, transition, states)

    return transition, state
end

Suppose we then wanted to use this with some of the packages which implements AbstractMCMC.jl's interface, e.g. AdvancedMH.jl, then we'd simply have to implement realize and realize!!:

function AbstractMCMC.updatestate!!(::AdvancedMH.Transition, state_prev::AdvancedMH.Transition)
    # Let's `deepcopy` just to be certain.
    return deepcopy(state_prev)
end

To use MixtureSampler with two samplers sampler1 and sampler2 from AdvancedMH.jl as components, we'd simply do

sampler = MixtureSampler([sampler1, sampler2], [0.1, 0.9])
transition, state = AbstractMCMC.step(rng, model, sampler)
while ...
    transition, state = AbstractMCMC.step(rng, model, sampler, state)
end

As a final note, there is one potential issue we haven't really addressed in the above implementation: a lot of samplers have their own implementations of AbstractMCMC.AbstractModel which means that we would also have to ensure that all the different samplers we are using would be compatible with the same model. A very easy way to fix this would be to just add a struct called ManyModels supporting getindex, e.g. models[i] would return the i-th model:

struct ManyModels{M} <: AbstractMCMC.AbstractModel
    models::M
end

Base.getindex(model::ManyModels, I...) = model.models[I...]

Then the above step would just extract the model corresponding to the current sampler:

# Take a `step` for this sampler using the updated state.
transition, state_current = AbstractMCMC.step(
    rng, model[i], sampler_current, state_current;
    kwargs...
)

This issue should eventually disappear as the community moves towards a unified approach to implement AbstractMCMC.AbstractModel.